backend/requirements.txt
boto3==1.35.68
botocore==1.35.96
fitz==0.0.1.dev2
openai==2.9.0
pydantic==2.12.5
pypdf==6.4.0
python-dotenv==1.2.1
quart==0.20.0
quart_cors==0.8.0
